requirements.txt
acres==0.2.0
aiofiles==24.1.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.1.0
frontend==0.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==2.1.3
os.path2==0.0.4
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
prov==2.0.1
puremagic==1.28
pydot==3.0.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
zarr==2.10.0
python-dotenv==0.19.2
pydicom==2.3.0

